Extract ShardRecoveryCancellation into a standalone class
