+++ server/src/test/java/org/elasticsearch/indices/recovery/CancelRecoveriesActionRequestTests.java
@@ -44,12 +44,12 @@ public class CancelRecoveriesActionRequestTests extends AbstractWireSerializingT
             );
     }
 
-    private List<CancelRecoveriesAction.ShardRecoveryCancellation> randomCancellations() {
+    private List<ShardRecoveryCancellation> randomCancellations() {
         final int size = randomIntBetween(0, 5);
-        final var cancellations = new ArrayList<CancelRecoveriesAction.ShardRecoveryCancellation>(size);
+        final var cancellations = new ArrayList<ShardRecoveryCancellation>(size);
         for (int i = 0; i < size; i++) {
             cancellations.add(
-                new CancelRecoveriesAction.ShardRecoveryCancellation(
+                new ShardRecoveryCancellation(
                     new ShardId(randomIdentifier(), UUIDs.randomBase64UUID(), i),
                     UUIDs.randomBase64UUID(),
                     randomBoolean()
